Add code to download matching OpenJDK source
This helps maintaining GPL compliance for OpenJDK much,
much easier, since the source code is being distributed
with the binaries.

I am downloading the source from GitHub because there is,
stupidly, no source tarball for OpenJDK 22.0.1 available on
the OpenJDK website. I have verified that this is the exact
build used in the binaries through consulting the
Contents/Home/release file in the downloaded JDK directories.